add implementation to handle when author to be deleted does not exist.

+++ src/main/java/com/bookstore/managementsystem/service/AuthorServiceImpl.java
@@ -83,8 +83,12 @@ public class AuthorServiceImpl implements AuthorService{
     }
 
     @Override
-    public ResponseEntity<Void> deleteAuthorById(Long id) {
+    public ResponseEntity<Void> deleteAuthorById(Long id) throws NotFoundError{
         boolean exists = authorRepo.existsById(id);
+
+        if(!exists) {
+            throw new NotFoundError("No Author found with the ID: "+id+".");
+        }
         authorRepo.deleteById(id);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
